+++ CHANGELOG.md
@@ -16,6 +16,7 @@ dev
  * ons-list-item: Fixed [#2150](https://github.com/OnsenUI/OnsenUI/issues/2150).
  * templates: Fixed a bug where `script` tags might not be ready on time.
  * ons-navigator: Fixed [#2167](https://github.com/OnsenUI/OnsenUI/issues/2167).
+ * ons-tabbar: Hide 1px line between toolbar and top tabbar in some Android devices.
 
 v2.5.3
 ----